Abstract servlets
